projects/proj_fastapi/pyproject.toml
[project]
name = "fanz-fastapi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.*",
  "uvicorn[standard]==0.30.*",
  "pydantic==2.*",
  "pydantic-settings==2.*",
  "SQLAlchemy[asyncio]==2.*",
  "asyncpg==0.29.*",
  "alembic==1.*",
  "python-multipart==0.0.*",
  "authlib==1.*",
  "argon2-cffi==23.*",
  "redis==5.*",
  "prometheus-client==0.20.*",
  "opentelemetry-sdk==1.*",
  "opentelemetry-exporter-otlp==1.*",
  "loguru==0.7.*"
]

[project.optional-dependencies]
dev = [
  "ruff==0.5.*",
  "black==24.*",
  "mypy==1.*",
  "pytest==8.*",
  "pytest-asyncio==0.23.*",
  "httpx==0.27.*",
  "pip-audit==2.*",
  "bandit==1.*"
]

[tool.ruff]
line-length = 100